move getDefaultChartType to helpers
