+++ src/helpers/getDefaultChartType.js
